Changed ratio type from Int to Double

+++ Ratios/Helpers/Calculator.swift
@@ -10,21 +10,21 @@ import Foundation
 
 struct Calculator {
 
-    static func calculateGrounds(brew: Double, ratio: Int) -> Double {
+    static func calculateGrounds(brew: Double, ratio: Double) -> Double {
         guard ratio > 2 else {
             return 0
         }
         return brew / Double(ratio - 2)
     }
 
-    static func calculateGrounds(water: Double, ratio: Int) -> Double {
+    static func calculateGrounds(water: Double, ratio: Double) -> Double {
         guard ratio > 0 else {
             return 0
         }
         return water / Double(ratio)
     }
 
-    static func calculateWater(grounds: Double, ratio: Int) -> Double {
+    static func calculateWater(grounds: Double, ratio: Double) -> Double {
         return grounds * Double(ratio)
     }
